SOS/requirements.txt
Django>=3.2,<4.0
mysqlclient>=2.0,<3.0
django-cors-headers
PyJWT>=2.0,<3.0